Refactored to only use one websocket connection
